+++ src/MutableCollection.php
@@ -28,6 +28,7 @@ use Somnambulist\Collection\Groups\Partitionable;
 use Somnambulist\Collection\Groups\Queryable;
 use Somnambulist\Collection\Groups\Runnable;
 use Somnambulist\Collection\Groups\StringHelpers;
+use Somnambulist\Collection\Utils\MapProxy;
 use Somnambulist\Collection\Utils\RunProxy;
 use Somnambulist\Collection\Utils\Value;
 
@@ -37,6 +38,7 @@ use Somnambulist\Collection\Utils\Value;
  * @package    Somnambulist\Collection
  * @subpackage Somnambulist\Collection\MutableCollection
  *
+ * @property-read MapProxy $map
  * @property-read RunProxy $run
  */
 class MutableCollection extends AbstractCollection implements
@@ -86,6 +88,9 @@ class MutableCollection extends AbstractCollection implements
     {
         if ('run' === $name && $this instanceof IsRunnable) {
             return new RunProxy($this);
+        }
+        if ('map' === $name && $this instanceof IsMappable) {
+            return new MapProxy($this);
         }
 
         return $this->offsetGet($name);