+++ src/MutableSet.php
@@ -29,6 +29,7 @@ use Somnambulist\Collection\Groups\Partitionable;
 use Somnambulist\Collection\Groups\Queryable;
 use Somnambulist\Collection\Groups\Runnable;
 use Somnambulist\Collection\Groups\StringHelpers;
+use Somnambulist\Collection\Utils\MapProxy;
 use Somnambulist\Collection\Utils\RunProxy;
 use Somnambulist\Collection\Utils\Value;
 
@@ -47,6 +48,7 @@ use Somnambulist\Collection\Utils\Value;
  * @package    Somnambulist\Collection
  * @subpackage Somnambulist\Collection\MutableSet
  *
+ * @property-read MapProxy $map
  * @property-read RunProxy $run
  */
 class MutableSet extends AbstractCollection implements
@@ -101,6 +103,9 @@ class MutableSet extends AbstractCollection implements
     {
         if ('run' === $name && $this instanceof IsRunnable) {
             return new RunProxy($this);
+        }
+        if ('map' === $name && $this instanceof IsMappable) {
+            return new MapProxy($this);
         }
 
         return $this->offsetGet($name);